+++ src/webview/js/Panel.js
@@ -1,6 +1,7 @@
 import { Animations } from './Animations';
 import { HierarchyTree } from './HierarchyTree';
 import { Info } from './Info';
+import { Materials } from './Materials';
 import { Textures } from './Textures';
 
 class Panel
@@ -16,6 +17,7 @@ class Panel
     this.contents = {
       hierarchy: new HierarchyTree(this, 'hierarchy'),
       textures: new Textures(this, 'textures'),
+      materials: new Materials(this, 'materials'),
       info: new Info(this, 'info'),
       animations: new Animations(this, 'animations')
     };
@@ -23,6 +25,7 @@ class Panel
     this.buttons = {
       hierarchy: this.$buttons_container.querySelector('.panel-button[data-name="hierarchy"]'),
       textures: this.$buttons_container.querySelector('.panel-button[data-name="textures"]'),
+      materials: this.$buttons_container.querySelector('.panel-button[data-name="materials"]'),
       info: this.$buttons_container.querySelector('.panel-button[data-name="info"]'),
       animations: this.$buttons_container.querySelector('.panel-button[data-name="animations"]')
     };
@@ -37,6 +40,7 @@ class Panel
   {
     this.contents.hierarchy.build_hierarchy_tree(object3d);
     this.contents.textures.update_contents(object3d);
+    this.contents.materials.update_contents(object3d);
   }
 
   handle_object_click(object3d)
@@ -73,6 +77,11 @@ class Panel
     if (this.contents.info.get_texture_count() > 0)
     {
       this.buttons.textures.classList.remove('hidden');
+    }
+
+    if (this.contents.materials.get_material_count() > 0)
+    {
+      this.buttons.materials.classList.remove('hidden');
     }
 
     if (this.contents.info.get_animation_count() > 0)