+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the GLB Viewer extension will be documented in this file.
 
+## [2.7.0] - 2025-08-02
+
+### Added
+ - Material window
+ - Material details window
+
 ## [2.6.0] - 2025-07-29
 
 ### Improved